admin comenzi filtre pt comenzi - rezolvat
modificare stergere comanda stoc- sa nu se poata sterge daca aceasta este livrata expediata in asteptare - rezolvat

Posibilitatea de a modifica cantitatea in cos si de a putea adauga mai multe produse de o data inaintea adaugarii in cos - rezolvat

TVA 19% global - admin temporar - rezzolvat

Admin vanzari totale pe o perioada stabilita de admin - rezolvat

Rezolvare cos sa se stearga dupa plasare comenzii - rezolvat

Login user si admin - daca ne logam cu un cont de admin sa fim redirectionati catre pagina admin - rezolvat

Cand apas logout de pe admin panel sa fiu trimis in pagina principala - rezolvat(trebuie sa dau logout de pe pagina principala)

Stergere redirectionare folosind iconite - rezolvat

Afisare coduri promotionale in admin promocode - rezolvat

Eroare daca adauga caractere la pret in adaugare produs - rezolvat

Profil user - toate comenziile, total comenzi TVA - rezolvat

Folosirea adresei salvate in checkout - rezolvat

Rezolvare user storage isadmin - rezolvat

Adauga posibilitatea de a da review-uri produsului - rezolvat

Adaugare descriere produs - rezolvat

Posibila problema la sortare de verificat dupa adaugarea produselor - rezolvat

+++ admin/src/Components/AfisareComenzi/AfisareComenzi.jsx
@@ -3,6 +3,16 @@ import './AfisareComenzi.css';
 
 const AfisareComenzi = () => {
   const [orders, setOrders] = useState([]);
+  const [filteredOrders, setFilteredOrders] = useState([]);
+  const [filters, setFilters] = useState({
+    orderId: '',
+    clientName: '',
+    email: '',
+    date: '',
+    product: '',
+    total: '',
+    status: '',
+  });
 
   useEffect(() => {
     const fetchOrders = async () => {
@@ -11,6 +21,7 @@ const AfisareComenzi = () => {
         const data = await response.json();
         if (data.success) {
           setOrders(data.orders);
+          setFilteredOrders(data.orders);
         } else {
           console.error('Eroare la preluarea comenzilor');
         }
@@ -21,6 +32,10 @@ const AfisareComenzi = () => {
 
     fetchOrders();
   }, []);
+
+  useEffect(() => {
+    filterOrders();
+  }, [filters, orders]);
 
   const updateOrderStatus = async (orderId, newStatus) => {
     try {
@@ -47,6 +62,13 @@ const AfisareComenzi = () => {
   };
 
   const handleDeleteOrder = async (orderId) => {
+    const order = orders.find(order => order._id === orderId);
+
+    if (order.status === 'Livrat' || order.status === 'Expediat' || order.status === 'În așteptare') {
+      console.error('Comanda nu poate fi ștearsă în starea actuală.');
+      return;
+    }
+
     try {
       const response = await fetch(`http://localhost:4000/deleteorder`, {
         method: 'DELETE',
@@ -71,6 +93,61 @@ const AfisareComenzi = () => {
     updateOrderStatus(orderId, 'Anulat');
   };
 
+  const handleFilterChange = (e) => {
+    setFilters({
+      ...filters,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const filterOrders = () => {
+    let filtered = orders;
+
+    if (filters.orderId) {
+      filtered = filtered.filter(order =>
+        order._id.toLowerCase().includes(filters.orderId.toLowerCase())
+      );
+    }
+
+    if (filters.clientName) {
+      filtered = filtered.filter(order =>
+        `${order.shippingDetails.firstName} ${order.shippingDetails.lastName}`
+          .toLowerCase()
+          .includes(filters.clientName.toLowerCase())
+      );
+    }
+
+    if (filters.email) {
+      filtered = filtered.filter(order =>
+        order.contactDetails.email.toLowerCase().includes(filters.email.toLowerCase())
+      );
+    }
+
+    if (filters.date) {
+      filtered = filtered.filter(order =>
+        new Date(order.date).toISOString().split('T')[0] === filters.date
+      );
+    }
+
+    if (filters.product) {
+      filtered = filtered.filter(order =>
+        order.cartItems.some(item =>
+          item.productName.toLowerCase().includes(filters.product.toLowerCase())
+        )
+      );
+    }
+
+    if (filters.total) {
+      filtered = filtered.filter(order => order.total.toString().includes(filters.total));
+    }
+
+    if (filters.status) {
+      filtered = filtered.filter(order => order.status.toLowerCase().includes(filters.status.toLowerCase()));
+    }
+
+    setFilteredOrders(filtered);
+  };
+
   return (
     <div className="orders-container">
       <h2 className="orders-header">Comenzi</h2>
@@ -78,18 +155,81 @@ const AfisareComenzi = () => {
         <table className="orders-table">
           <thead>
             <tr>
-              <th>ID Comandă</th>
-              <th>Nume Client</th>
-              <th>Email</th>
-              <th>Data</th>
-              <th>Produse</th>
-              <th>Total</th>
-              <th>Status</th>
+              <th>
+                ID Comandă
+                <input
+                  type="text"
+                  name="orderId"
+                  placeholder="ID Comandă"
+                  value={filters.orderId}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Nume Client
+                <input
+                  type="text"
+                  name="clientName"
+                  placeholder="Nume Client"
+                  value={filters.clientName}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Email
+                <input
+                  type="email"
+                  name="email"
+                  placeholder="Email"
+                  value={filters.email}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Data
+                <input
+                  type="date"
+                  name="date"
+                  placeholder="Data"
+                  value={filters.date}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Produse
+                <input
+                  type="text"
+                  name="product"
+                  placeholder="Produse"
+                  value={filters.product}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Total
+                <input
+                  type="text"
+                  name="total"
+                  placeholder="Total"
+                  value={filters.total}
+                  onChange={handleFilterChange}
+                />
+              </th>
+              <th>
+                Status
+                <input
+                  type="text"
+                  name="status"
+                  placeholder="Status"
+                  value={filters.status}
+                  onChange={handleFilterChange}
+                />
+              </th>
               <th>Acțiuni</th>
             </tr>
           </thead>
           <tbody>
-            {orders.map((order, index) => (
+            {filteredOrders.map((order, index) => (
               <tr key={index}>
                 <td data-label="ID Comandă">{order._id}</td>
                 <td data-label="Nume Client">{order.shippingDetails.firstName} {order.shippingDetails.lastName}</td>
@@ -107,7 +247,9 @@ const AfisareComenzi = () => {
                 <td data-label="Total">{order.total} RON</td>
                 <td data-label="Status">{order.status}</td>
                 <td data-label="Acțiuni">
-                  <button onClick={() => handleDeleteOrder(order._id)}>Sterge Comanda</button>
+                  <button onClick={() => handleDeleteOrder(order._id)} disabled={order.status === 'Livrat' || order.status === 'Expediat' || order.status === 'În așteptare'}>
+                    Șterge Comanda
+                  </button>
                   <button 
                     onClick={() => handleCancelOrder(order._id)} 
                     disabled={order.status === 'Anulat'}
